refactor: centralize contract schema
